load_tests/requirements.txt
locust>=2.17.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
websocket-client>=1.6.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
